can filter

+++ common/src/infrastructure/db/sqlx/flower.rs
@@ -1,3 +1,4 @@
+
 use common_type::models::flowers::Flower as ModelFlower;
 
 #[derive(sqlx::FromRow, sqlx::Type, Debug, Clone)]
@@ -65,7 +66,7 @@ pub struct Flower {
 impl Flower {
     pub fn to_model_flower(&self) -> ModelFlower {
         ModelFlower {
-            id: self.id.map(|l| l as i128),
+            id: self.id.map(|l| l as f64),
             name_jp: self.name_jp.clone(),
             name_en: self.name_en.clone(),
             scientific_name: self.scientific_name.clone(),
